+++ src/pybind/mgr/tox.ini
@@ -1,15 +1,22 @@
 [tox]
 envlist = py3, mypy
 skipsdist = true
+requires = cython
 
 [testenv]
-setenv = UNITTEST = true
-deps = -r requirements.txt
+setenv =
+    UNITTEST = true
+    PYTHONPATH = ../../../build/lib/cython_modules/lib.3/
+    LD_LIBRARY_PATH = ../../../build/lib
+deps =
+    cython
+    -r requirements.txt
 commands = pytest -v --cov --cov-append --cov-report=term --doctest-modules {posargs:mgr_util.py tests/ cephadm/ pg_autoscaler/ progress/}
 
 [testenv:mypy]
 basepython = python3
 deps =
+    cython
     -r requirements.txt
     mypy
 commands = mypy --config-file=../../mypy.ini \
